Fix Swing Demo Frame
