+++ .gitignore
@@ -1,6 +1,3 @@
-/root/object-criteria-factory/target/
-/root/entity-mapper/target/
-/root/el-mapper-example/target/
 
 .vscode
 target
