+++ entity-mapper-demo/pom.xml
@@ -82,7 +82,11 @@
             <version>1.0</version>
             <type>jar</type>
         </dependency>
-
+        <dependency>
+            <groupId>org.netbeans.external</groupId>
+            <artifactId>AbsoluteLayout</artifactId>
+            <version>RELEASE140</version>
+        </dependency>
     </dependencies>
 
 </project>
